Create bfs level node.cpp
You have been given a Tree consisting of N nodes. A tree is a fully-connected graph consisting of N nodes and  N-1 edges. The nodes in this tree are indexed from 1 to N. Consider node indexed 1 to be the root node of this tree. The root node lies at level one in the tree. You shall be given the tree and a single integer x . You need to find out the number of nodes lying on level x.